[x86_64] Implement GDT

+++ kernel/kernel/segmentation/gdt.h
@@ -5,6 +5,6 @@
 
 void gdt_init(void);
 // Must be performed upon task switch to allow the kernel to be preemptible
-void tss_set_kernel_stack(uint32_t stack);
+void tss_set_kernel_stack(uint64_t stack);
 
 #endif
